refactor(transaction-controller): rename Revert types, simplify revert util
- Rename type `RevertSources` \u2192 `RevertData` (parent shape on TransactionMeta).
- Rename file `extract-revert-reason.ts` \u2192 `revert.ts`.
- New single public `decodeRevert(data)` accepting raw hex; returns `Revert`
  with both decoded `message` and the original `data`.
- `extractRevert` swallows all errors internally and returns `undefined`.
- Move simulation revert write inside the `isBalanceChangesSkipped` check
  so `revert.simulation` and `simulationData` stay in sync.
- Single `metamask:transaction-controller:revert` debug logger; drop the
  earlier `extract-revert-reason` namespace.
- Trim `TransactionMeta.revert` JSDoc to one line; details live on the
  per-source `Revert` and on `RevertData`.
- Link Solidity panic spec from `PANIC_CODE_MESSAGES` JSDoc; codes verified
  against https://docs.soliditylang.org/en/latest/control-structures.html#panic-via-assert-and-error-via-require.
